+++ CloudHarvestCLI/commands/__init__.py
@@ -2,3 +2,4 @@ from CloudHarvestCLI.commands.banner.commandset import BannerCommand
 from CloudHarvestCLI.commands.cache.commandset import CacheCommand
 from CloudHarvestCLI.commands.report.commandset import ReportCommand
 from CloudHarvestCLI.commands.services.commandset import ServicesCommand
+from CloudHarvestCLI.commands.theme.commandset import ThemeCommand